Devided components into each files
