+++ tic-tac-toe-app/src/Square.tsx
@@ -1,0 +1,12 @@
+export const Square = ({ value, onSquareClick }: SquareProps) => {
+  return (
+    <button className="square" onClick={onSquareClick}>
+      {value}
+    </button>
+  );
+};
+
+type SquareProps = {
+  value: string;
+  onSquareClick: () => void;
+};
