+++ tic-tac-toe-app/src/getUser.ts
@@ -1,0 +1,1 @@
+export const getUser = (b: boolean): string => (b ? 'X' : 'O');
